+++ src/frontend/OverlayManager/Overlay/NavBarOverlayWrapper/index.module.scss
@@ -1,0 +1,3 @@
+.disabled {
+  cursor: not-allowed !important;
+}
