+++ src/frontend/OverlayManager/Overlay/index.module.scss
@@ -56,7 +56,3 @@
 .hideOverlay {
   display: none;
 }
-
-.disabled {
-  cursor: not-allowed !important;
-}
