+++ src/frontend/screens/MetaMaskPortfolio/index.module.css
@@ -1,5 +1,6 @@
 .homeDiv {
   height: 100%;
+  width: 100%;
 }
 
 .homeWebview {
